feat: Create module bootstrap
