--- toolbox/modal.js
@@ -1,94 +1,0 @@
-import { Component } from "../core/components.js";
-
-export class ModalComponent extends Component {
-    constructor() {
-        super({
-            __active: false
-        });
-        this.onOpen = this.onClose = () => null;
-    }
-
-    onCreate() {
-        const onOpen = this.getFunctionAttribute("onOpen", "event");
-        const onClose = this.getFunctionAttribute("onClose", "event");
-
-        if (onOpen) {
-            this.onOpen = onOpen;
-        }
-        if (onClose) {
-            this.onClose = onClose;
-        }
-    }
-
-    setOnOpen(callback) {
-        this.onOpen = callback;
-    }
-
-    setOnClose(callback) {
-        this.onClose = callback;
-    }
-
-    register(object) {
-        for (const key in object) {
-            this[key] = object[key];
-        }
-    }
-
-    open() {
-        this.__active = true;
-        this.onOpen();
-    }
-
-    close() {
-        this.__active = false;
-        this.onClose();
-    }
-
-    style() {
-        return `& .modal-open {overflow: hidden;} & .modal-open .modal {overflow-x: hidden;overflow-y: auto;} & .modal {position: fixed;top: 0;left: 0;z-index: 1055;width: 100%;height: 100%;overflow-x: hidden;overflow-y: auto;outline: 0;} & .modal:not(.show) {display: none;} & .modal.show {display: block;} & .modal-dialog {position: relative;width: auto;margin: 0.5rem;pointer-events: none;} & .modal.fade .modal-dialog {transition: transform 0.3s ease-out;transform: translate(0, -50px);}@media (prefers-reduced-motion: reduce) {& .modal.fade .modal-dialog {transition: none;}} & .modal.show .modal-dialog {transform: none;} & .modal.modal-static .modal-dialog {transform: scale(1.02);} & .modal-dialog-scrollable {height: calc(100% - 1rem);} & .modal-dialog-scrollable .modal-content {max-height: 100%;overflow: hidden;} & .modal-dialog-scrollable .modal-body {overflow-y: auto;} & .modal-dialog-centered {display: flex;align-items: center;min-height: calc(100% - 1rem);} & .modal-content {position: relative;display: flex;flex-direction: column;width: 100%;pointer-events: auto;background-color: #fff;background-clip: padding-box;border: 1px solid rgba(0, 0, 0, 0.2);border-radius: 0.3rem;outline: 0;} & .modal-backdrop {position: fixed;top: 0;left: 0;z-index: 1050;width: 100vw;height: 100vh;background-color: #000;} & .modal-backdrop.fade {opacity: 0;} & .modal-backdrop.show {opacity: 0.5;} & .modal-header {display: flex;flex-shrink: 0;align-items: center;justify-content: space-between;padding: 1rem 1rem;border-bottom: 1px solid #dee2e6;border-top-left-radius: calc(0.3rem - 1px);border-top-right-radius: calc(0.3rem - 1px);} & .modal-header .btn-close {padding: 0.5rem 0.5rem;margin: -0.5rem -0.5rem -0.5rem auto;} & .modal-title {margin-bottom: 0;line-height: 1.5;} & .modal-body {position: relative;flex: 1 1 auto;padding: 1rem;} & .modal-footer {display: flex;flex-wrap: wrap;flex-shrink: 0;align-items: center;justify-content: flex-end;padding: 0.75rem;border-top: 1px solid #dee2e6;border-bottom-right-radius: calc(0.3rem - 1px);border-bottom-left-radius: calc(0.3rem - 1px);} & .modal-footer > * {margin: 0.25rem;}@media (min-width: 576px) {& .modal-dialog {max-width: 500px;margin: 1.75rem auto;} & .modal-dialog-scrollable {height: calc(100% - 3.5rem);} & .modal-dialog-centered {min-height: calc(100% - 3.5rem);} & .modal-sm {max-width: 300px;}}@media (min-width: 992px) {& .modal-lg, & .modal-xl {max-width: 800px;}}@media (min-width: 1200px) {& .modal-xl {max-width: 1140px;}} & .modal-fullscreen {width: 100vw;max-width: none;height: 100%;margin: 0;} & .modal-fullscreen .modal-content {height: 100%;border: 0;border-radius: 0;} & .modal-fullscreen .modal-header {border-radius: 0;} & .modal-fullscreen .modal-body {overflow-y: auto;} & .modal-fullscreen .modal-footer {border-radius: 0;}@media (max-width: 575.98px) {& .modal-fullscreen-sm-down {width: 100vw;max-width: none;height: 100%;margin: 0;} & .modal-fullscreen-sm-down .modal-content {height: 100%;border: 0;border-radius: 0;} & .modal-fullscreen-sm-down .modal-header {border-radius: 0;} & .modal-fullscreen-sm-down .modal-body {overflow-y: auto;} & .modal-fullscreen-sm-down .modal-footer {border-radius: 0;}}@media (max-width: 767.98px) {& .modal-fullscreen-md-down {width: 100vw;max-width: none;height: 100%;margin: 0;} & .modal-fullscreen-md-down .modal-content {height: 100%;border: 0;border-radius: 0;} & .modal-fullscreen-md-down .modal-header {border-radius: 0;} & .modal-fullscreen-md-down .modal-body {overflow-y: auto;} & .modal-fullscreen-md-down .modal-footer {border-radius: 0;}}@media (max-width: 991.98px) {& .modal-fullscreen-lg-down {width: 100vw;max-width: none;height: 100%;margin: 0;} & .modal-fullscreen-lg-down .modal-content {height: 100%;border: 0;border-radius: 0;} & .modal-fullscreen-lg-down .modal-header {border-radius: 0;} & .modal-fullscreen-lg-down .modal-body {overflow-y: auto;} & .modal-fullscreen-lg-down .modal-footer {border-radius: 0;}}@media (max-width: 1199.98px) {& .modal-fullscreen-xl-down {width: 100vw;max-width: none;height: 100%;margin: 0;} & .modal-fullscreen-xl-down .modal-content {height: 100%;border: 0;border-radius: 0;} & .modal-fullscreen-xl-down .modal-header {border-radius: 0;} & .modal-fullscreen-xl-down .modal-body {overflow-y: auto;} & .modal-fullscreen-xl-down .modal-footer {border-radius: 0;}}@media (max-width: 1399.98px) {& .modal-fullscreen-xxl-down {width: 100vw;max-width: none;height: 100%;margin: 0;} & .modal-fullscreen-xxl-down .modal-content {height: 100%;border: 0;border-radius: 0;} & .modal-fullscreen-xxl-down .modal-header {border-radius: 0;} & .modal-fullscreen-xxl-down .modal-body {overflow-y: auto;} & .modal-fullscreen-xxl-down .modal-footer {border-radius: 0;}}`;
-    }
-
-    render() {
-        return `<div class="modal ${this.getAttribute("modal-class") || ''} ${this.__active ? 'show' : ''}" tabindex="-1" role="dialog" aria-hidden="true"><div class="modal-dialog ${this.getAttribute("modal-dialog-class") || ''}" role="document"><div class="modal-content ${this.getAttribute("modal-content-class") || ''}">${this.children.innerHTML}</div></div></div>${this.__active?`<div class="modal-backdrop fade show"></div>`: ""}`;
-    }
-}
-
-export class ModalContainer extends Component {
-    constructor(componentClass, ...args) {
-        super();
-        this.modal = new ModalComponent();
-        this.content = new componentClass(this, ...args);
-        this.appendChild(this.modal, "modal-component");
-        this.modal.appendChild(this.content, "modal-content");
-    }
-
-    open() {
-        this.modal.open();
-    }
-
-    close() {
-        this.modal.close();
-    }
-
-    setOnOpen(callback) {
-        this.modal.setOnOpen(callback);
-    }
-
-    setOnClose(callback) {
-        this.modal.setOnClose(callback);
-    }
-
-    register(fields) {
-        for (const key in fields) {
-            this[key] = fields[key];
-        }
-    }
-
-    getContent() {
-        return this.content;
-    }
-
-    render() {
-        return `<modal-component><modal-content></modal-content></modal-component>`;
-    }
-}
